test(login,mockapi): Add tests, modify database models and update dependencies
In this commit, unit tests for mockapi serializers and login views are added to ensure proper functionality. Database model changes include addition of new columns and alteration of existing columns in 'MockAPI' model. Project dependencies are also updated with the addition of pytest. Last but not the least, login views have been updated for returning appropriate HTTP status codes and URL configurations updated.