Test/login mockapi test (#209)
* test(login,mockapi): Add tests, modify database models and update dependencies

In this commit, unit tests for mockapi serializers and login views are added to ensure proper functionality. Database model changes include addition of new columns and alteration of existing columns in 'MockAPI' model. Project dependencies are also updated with the addition of pytest. Last but not the least, login views have been updated for returning appropriate HTTP status codes and URL configurations updated.

* ci: Add GitHub Actions workflow and alter Django model

A GitHub Actions workflow for running pytest and coverage has been added to the project. In addition, a modification to the Django model has been made to alter the "first_name" field for the "MyUser" model.

* feat(mock): add MockAPIResponseHandler model

This commit adds a new model called `MockAPIResponseHandler` to the `mock` app. The model includes fields for creating and updating time, creator and updater information, condition expression, response set value expression, response set value, enabled status, and a foreign key to the `MockAPI` model. This model will be used to handle responses for mock API requests.

feat(mock): add description and name fields to MockAPIResponseHandler

This commit adds two new fields, `description` and `name`, to the `MockAPIResponseHandler` model in the `mock` app. The `description` field is a text field for providing a description of the response handler, and the `name` field is a character field with an index for storing the name of the response handler.

---------

Co-authored-by: rikasai233 <rikasai233@gmail.com>